4.1.0

+++ src/main/java/net/fellter/vanillablocksplus/custom_blocks/slime/SlimeFenceBlock.java
@@ -16,7 +16,7 @@ public class SlimeFenceBlock extends FenceBlock {
     }
 
     @Override
-    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
         if (entity.bypassesLandingEffects()) {
             super.onLandedUpon(world, state, pos, entity, fallDistance);
         } else {
