+++ src/main/java/net/fellter/vanillablocksplus/custom_blocks/slime/SlimePressurePlateBlock.java
@@ -17,7 +17,7 @@ public class SlimePressurePlateBlock extends PressurePlateBlock {
     }
 
     @Override
-    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+    public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
         if (entity.bypassesLandingEffects()) {
             super.onLandedUpon(world, state, pos, entity, fallDistance);
         } else {
